add ExtractorTests and make ConstantNullBlock implement AggregateMetricDoubleBlock
