+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/AggregateMetricDoubleBlock.java
@@ -7,297 +7,99 @@
 
 package org.elasticsearch.compute.data;
 
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.core.ReleasableIterator;
-import org.elasticsearch.core.Releasables;
 
-import java.io.IOException;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Stream;
 
-public final class AggregateMetricDoubleBlock extends AbstractNonThreadSafeRefCounted implements Block {
-    private final DoubleBlock minBlock;
-    private final DoubleBlock maxBlock;
-    private final DoubleBlock sumBlock;
-    private final IntBlock countBlock;
-    private final int positionCount;
+/**
+ * Block that stores aggregate_metric_double values.
+ */
+public sealed interface AggregateMetricDoubleBlock extends Block permits AggregateMetricDoubleArrayBlock, ConstantNullBlock {
 
-    public AggregateMetricDoubleBlock(DoubleBlock minBlock, DoubleBlock maxBlock, DoubleBlock sumBlock, IntBlock countBlock) {
-        this.minBlock = minBlock;
-        this.maxBlock = maxBlock;
-        this.sumBlock = sumBlock;
-        this.countBlock = countBlock;
-        this.positionCount = minBlock.getPositionCount();
-        for (Block b : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            if (b.getPositionCount() != positionCount) {
-                assert false : "expected positionCount=" + positionCount + " but was " + b;
-                throw new IllegalArgumentException("expected positionCount=" + positionCount + " but was " + b);
-            }
-            if (b.isReleased()) {
-                assert false : "can't build aggregate_metric_double block out of released blocks but [" + b + "] was released";
-                throw new IllegalArgumentException(
-                    "can't build aggregate_metric_double block out of released blocks but [" + b + "] was released"
-                );
-            }
+    @Override
+    AggregateMetricDoubleBlock filter(int... positions);
+
+    @Override
+    AggregateMetricDoubleBlock keepMask(BooleanVector mask);
+
+    @Override
+    ReleasableIterator<? extends AggregateMetricDoubleBlock> lookup(IntBlock positions, ByteSizeValue targetBlockSize);
+
+    @Override
+    AggregateMetricDoubleBlock expand();
+
+    /**
+     * Compares the given object with this block for equality. Returns {@code true} if and only if the
+     * given object is a AggregateMetricDoubleBlock, and both blocks are
+     * {@link #equals(AggregateMetricDoubleBlock, AggregateMetricDoubleBlock) equal}.
+     */
+    @Override
+    boolean equals(Object obj);
+
+    /** Returns the hash code of this block, as defined by {@link #hash(AggregateMetricDoubleBlock)}. */
+    @Override
+    int hashCode();
+
+    /**
+     * Returns {@code true} if the given blocks are equal to each other, otherwise {@code false}.
+     * Two blocks are considered equal if they have the same position count, and contain the same
+     * values (including absent null values) in the same order. This definition ensures that the
+     * equals method works properly across different implementations of the AggregateMetricDoubleBlock interface.
+     */
+    static boolean equals(AggregateMetricDoubleBlock block1, AggregateMetricDoubleBlock block2) {
+        if (block1 == block2) {
+            return true;
         }
-    }
-
-    public static AggregateMetricDoubleBlock fromCompositeBlock(CompositeBlock block) {
-        assert block.getBlockCount() == 4
-            : "Can't make AggregateMetricDoubleBlock out of CompositeBlock with " + block.getBlockCount() + " blocks";
-        DoubleBlock min = block.getBlock(AggregateMetricDoubleBlockBuilder.Metric.MIN.getIndex());
-        DoubleBlock max = block.getBlock(AggregateMetricDoubleBlockBuilder.Metric.MAX.getIndex());
-        DoubleBlock sum = block.getBlock(AggregateMetricDoubleBlockBuilder.Metric.SUM.getIndex());
-        IntBlock count = block.getBlock(AggregateMetricDoubleBlockBuilder.Metric.COUNT.getIndex());
-        return new AggregateMetricDoubleBlock(min, max, sum, count);
-    }
-
-    public CompositeBlock asCompositeBlock() {
-        final Block[] blocks = new Block[4];
-        blocks[AggregateMetricDoubleBlockBuilder.Metric.MIN.getIndex()] = minBlock;
-        blocks[AggregateMetricDoubleBlockBuilder.Metric.MAX.getIndex()] = maxBlock;
-        blocks[AggregateMetricDoubleBlockBuilder.Metric.SUM.getIndex()] = sumBlock;
-        blocks[AggregateMetricDoubleBlockBuilder.Metric.COUNT.getIndex()] = countBlock;
-        return new CompositeBlock(blocks);
-    }
-
-    @Override
-    protected void closeInternal() {
-        Releasables.close(minBlock, maxBlock, sumBlock, countBlock);
-    }
-
-    @Override
-    public Vector asVector() {
-        return null;
-    }
-
-    @Override
-    public int getTotalValueCount() {
-        int totalValueCount = 0;
-        for (Block b : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            totalValueCount += b.getTotalValueCount();
+        final int positions = block1.getPositionCount();
+        if (positions != block2.getPositionCount()) {
+            return false;
         }
-        return totalValueCount;
-    }
-
-    @Override
-    public int getPositionCount() {
-        return positionCount;
-    }
-
-    @Override
-    public int getFirstValueIndex(int position) {
-        return minBlock.getFirstValueIndex(position);
-    }
-
-    @Override
-    public int getValueCount(int position) {
-        int max = 0;
-        for (Block b : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            max = Math.max(max, b.getValueCount(position));
-        }
-        return max;
-    }
-
-    @Override
-    public ElementType elementType() {
-        return ElementType.AGGREGATE_METRIC_DOUBLE;
-    }
-
-    @Override
-    public BlockFactory blockFactory() {
-        return minBlock.blockFactory();
-    }
-
-    @Override
-    public void allowPassingToDifferentDriver() {
-        for (Block block : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            block.allowPassingToDifferentDriver();
-        }
-    }
-
-    @Override
-    public boolean isNull(int position) {
-        for (Block block : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            if (block.isNull(position) == false) {
-                return false;
+        for (int pos = 0; pos < positions; pos++) {
+            if (block1.isNull(pos) || block2.isNull(pos)) {
+                if (block1.isNull(pos) != block2.isNull(pos)) {
+                    return false;
+                }
+            } else {
+                final int valueCount = block1.getValueCount(pos);
+                if (valueCount != block2.getValueCount(pos)) {
+                    return false;
+                }
+                for (var doubleMetric : List.of(
+                    AggregateMetricDoubleBlockBuilder.Metric.MIN,
+                    AggregateMetricDoubleBlockBuilder.Metric.MAX,
+                    AggregateMetricDoubleBlockBuilder.Metric.SUM
+                )) {
+                    DoubleBlock doubleBlock1 = (DoubleBlock) block1.getMetricBlock(doubleMetric.getIndex());
+                    DoubleBlock doubleBlock2 = (DoubleBlock) block2.getMetricBlock(doubleMetric.getIndex());
+                    if (DoubleBlock.equals(doubleBlock1, doubleBlock2) == false) {
+                        return false;
+                    }
+                }
+                IntBlock intBlock1 = block1.countBlock();
+                IntBlock intBlock2 = block2.countBlock();
+                return IntBlock.equals(intBlock1, intBlock2);
             }
         }
         return true;
     }
 
-    @Override
-    public boolean mayHaveNulls() {
-        return Stream.of(minBlock, maxBlock, sumBlock, countBlock).anyMatch(Block::mayHaveNulls);
+    static int hash(AggregateMetricDoubleBlock block) {
+        int result = 1;
+        result = 31 * result + DoubleBlock.hash(block.minBlock());
+        result = 31 * result + DoubleBlock.hash(block.maxBlock());
+        result = 31 * result + DoubleBlock.hash(block.sumBlock());
+        result = 31 * result + IntBlock.hash(block.countBlock());
+        return result;
     }
 
-    @Override
-    public boolean areAllValuesNull() {
-        return Stream.of(minBlock, maxBlock, sumBlock, countBlock).allMatch(Block::areAllValuesNull);
-    }
+    DoubleBlock minBlock();
 
-    @Override
-    public boolean mayHaveMultivaluedFields() {
-        return Stream.of(minBlock, maxBlock, sumBlock, countBlock).anyMatch(Block::mayHaveMultivaluedFields);
-    }
+    DoubleBlock maxBlock();
 
-    @Override
-    public boolean doesHaveMultivaluedFields() {
-        if (Stream.of(minBlock, maxBlock, sumBlock, countBlock).noneMatch(Block::mayHaveMultivaluedFields)) {
-            return false;
-        }
-        return Stream.of(minBlock, maxBlock, sumBlock, countBlock).anyMatch(Block::doesHaveMultivaluedFields);
-    }
+    DoubleBlock sumBlock();
 
-    @Override
-    public Block filter(int... positions) {
-        AggregateMetricDoubleBlock result = null;
-        DoubleBlock newMinBlock = null;
-        DoubleBlock newMaxBlock = null;
-        DoubleBlock newSumBlock = null;
-        IntBlock newCountBlock = null;
-        try {
-            newMinBlock = minBlock.filter(positions);
-            newMaxBlock = maxBlock.filter(positions);
-            newSumBlock = sumBlock.filter(positions);
-            newCountBlock = countBlock.filter(positions);
-            result = new AggregateMetricDoubleBlock(newMinBlock, newMaxBlock, newSumBlock, newCountBlock);
-            return result;
-        } finally {
-            if (result == null) {
-                Releasables.close(newMinBlock, newMaxBlock, newSumBlock, newCountBlock);
-            }
-        }
-    }
+    IntBlock countBlock();
 
-    @Override
-    public Block keepMask(BooleanVector mask) {
-        AggregateMetricDoubleBlock result = null;
-        DoubleBlock newMinBlock = null;
-        DoubleBlock newMaxBlock = null;
-        DoubleBlock newSumBlock = null;
-        IntBlock newCountBlock = null;
-        try {
-            newMinBlock = minBlock.keepMask(mask);
-            newMaxBlock = maxBlock.keepMask(mask);
-            newSumBlock = sumBlock.keepMask(mask);
-            newCountBlock = countBlock.keepMask(mask);
-            result = new AggregateMetricDoubleBlock(newMinBlock, newMaxBlock, newSumBlock, newCountBlock);
-            return result;
-        } finally {
-            if (result == null) {
-                Releasables.close(newMinBlock, newMaxBlock, newSumBlock, newCountBlock);
-            }
-        }
-    }
-
-    @Override
-    public ReleasableIterator<? extends Block> lookup(IntBlock positions, ByteSizeValue targetBlockSize) {
-        // TODO: support
-        throw new UnsupportedOperationException("can't lookup values from AggregateMetricDoubleBlock");
-    }
-
-    @Override
-    public MvOrdering mvOrdering() {
-        // TODO: determine based on sub-blocks
-        return MvOrdering.UNORDERED;
-    }
-
-    @Override
-    public Block expand() {
-        this.incRef();
-        return this;
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        for (Block block : List.of(minBlock, maxBlock, sumBlock, countBlock)) {
-            block.writeTo(out);
-        }
-    }
-
-    public static Block readFrom(StreamInput in) throws IOException {
-        boolean success = false;
-        DoubleBlock minBlock = null;
-        DoubleBlock maxBlock = null;
-        DoubleBlock sumBlock = null;
-        IntBlock countBlock = null;
-        BlockStreamInput blockStreamInput = (BlockStreamInput) in;
-        try {
-            minBlock = DoubleBlock.readFrom(blockStreamInput);
-            maxBlock = DoubleBlock.readFrom(blockStreamInput);
-            sumBlock = DoubleBlock.readFrom(blockStreamInput);
-            countBlock = IntBlock.readFrom(blockStreamInput);
-            AggregateMetricDoubleBlock result = new AggregateMetricDoubleBlock(minBlock, maxBlock, sumBlock, countBlock);
-            success = true;
-            return result;
-        } finally {
-            if (success == false) {
-                Releasables.close(minBlock, maxBlock, sumBlock, countBlock);
-            }
-        }
-    }
-
-    @Override
-    public long ramBytesUsed() {
-        return minBlock.ramBytesUsed() + maxBlock.ramBytesUsed() + sumBlock.ramBytesUsed() + countBlock.ramBytesUsed();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        AggregateMetricDoubleBlock that = (AggregateMetricDoubleBlock) o;
-        return positionCount == that.positionCount
-            && minBlock.equals(that.minBlock)
-            && maxBlock.equals(that.maxBlock)
-            && sumBlock.equals(that.sumBlock)
-            && countBlock.equals(that.countBlock);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(
-            DoubleBlock.hash(minBlock),
-            DoubleBlock.hash(maxBlock),
-            DoubleBlock.hash(sumBlock),
-            IntBlock.hash(countBlock),
-            positionCount
-        );
-    }
-
-    public DoubleBlock minBlock() {
-        return minBlock;
-    }
-
-    public DoubleBlock maxBlock() {
-        return maxBlock;
-    }
-
-    public DoubleBlock sumBlock() {
-        return sumBlock;
-    }
-
-    public IntBlock countBlock() {
-        return countBlock;
-    }
-
-    public Block getMetricBlock(int index) {
-        if (index == AggregateMetricDoubleBlockBuilder.Metric.MIN.getIndex()) {
-            return minBlock;
-        }
-        if (index == AggregateMetricDoubleBlockBuilder.Metric.MAX.getIndex()) {
-            return maxBlock;
-        }
-        if (index == AggregateMetricDoubleBlockBuilder.Metric.SUM.getIndex()) {
-            return sumBlock;
-        }
-        if (index == AggregateMetricDoubleBlockBuilder.Metric.COUNT.getIndex()) {
-            return countBlock;
-        }
-        throw new UnsupportedOperationException("Received an index (" + index + ") outside of range for AggregateMetricDoubleBlock.");
-    }
+    Block getMetricBlock(int index);
 }
